Music, Better controls and an attempt at implementing a dash control.
I added a background music that loops in the background while you play. It's actually pretty annoying so I'll probably change it. I also made it so that you can move in mid air as I hated the default controls in the tutorial. Finally, I have attempted to make a dashing function that has a cooldown but I'm too dumb to figure it out yet.